Fixed moving befor symlinking to only move files not symlinks...

+++ Rakefile
@@ -13,7 +13,7 @@ task :create_symlinks do
 	symlinks = Dir.glob("#{Dir.pwd}/*.symlink")
 	symlinks.each do |symlink|
     dest_filename = "#{Dir.home}/.#{File.basename(symlink).sub('.symlink', '')}"
-    if File.exist?(dest_filename) then
+    if File.exist?(dest_filename) and !File.symlink?(dest_filename) then
       FileUtils.mv(dest_filename, "#{dest_filename}.old")
     end
 		FileUtils.ln_sf(symlink, "#{Dir.home}/.#{File.basename(symlink).sub('.symlink', '')}" )
